init commit; mid flight update with config dataclasses still needed for metrics and activations
