+++ configs/1702.yaml
@@ -1,0 +1,23 @@
+wandb
+  project_name: 'fronts'
+  model_run_name: '1702_retrain'
+  # -- Optional parameters -- 
+  # log_frequency: 
+  # upload_checkpoints: 
+  # api_key: 
+  # wandb_filepath:
+
+train:
+  activation: gelu
+  activity_regularizer: 
+  batch_normalization: True
+  batch_sizes: 64
+  bias_constraint:
+  bias_initializer: zeros
+  bias_regularizer:
+  classes: 6
+  learning_rate: 0.0001
+
+validation:
+  batch_size: 64
+
